refactor: 메인페이지에서 사용되는 커뮤니티 섹션의 컴포넌트 이름 변경
- 실제 커뮤니티 페이지의 컴포넌트와 구분이 쉽도록
