+++ src/components/CommunitySection/IndexCommunityBanner.tsx
@@ -9,7 +9,7 @@ import Speaker from '@images/speaker.svg?react';
 
 import styles from './index.module.css';
 
-export const CommunityBanner: React.FC<PropH> = ({ h }) => (
+export const IndexCommunityBanner: React.FC<PropH> = ({ h }) => (
   <Flex
     bg={BoomerangColors.deepBlue}
     h={h}
